refactor(vscode): lift focus inspector out of behavior.ts
Slice of #767. Extracts the side panel that appears in focus mode —
Inspect product picker (a11y / Layout / Strings / Resources / Fonts /
Render / Theme / Recomposition), History accordion (diff vs HEAD,
diff vs main), Tools strip (a11y, device, live, record action
buttons) — into a typed `FocusInspectorController`.

The eight helpers (`sectionHeader` / `actionButton` / `historyPanel`
/ `productSpec` / `productPicker` / `productOption` /
`toggleFocusProduct` / `buildFocusPlaceholders`) move into the same
module as private functions / methods. The persistent UI state
(`enabledFocusProducts` Set, `focusProductPickerOpen` /
`focusHistoryOpen` flags) becomes private state on the controller.

Action callbacks (`toggleA11yOverlay` / `toggleInteractive` /
`toggleRecording` / `requestFocusedDiff` / `requestLaunchOnDevice`)
and per-preview state queries (a11y findings / nodes, live set,
store-backed `a11yOverlayId`) come in via `FocusInspectorConfig`. The
`clearAll` message path now calls `inspector.clearProducts()` to
reset the Set.

Drops `renderSummary` and `themeSummary` from the imperative copy —
both were dead code (defined but never called).

`behavior.ts` shrinks from 2015 to 1727 lines (-288 LOC).

Refs #767.
